engine: Fix anaconda version problem
If the version of the kernel/initrd/squashfs used for installation is to
far ahead of the os version beeing installed, it can lead to problems.

The reason is that anaconda - during installatin - is writing a
"flattened" kickstart (used for the installation) to disk. This
kickstart is then used by initial-setup.
If the kickstart is to new and contains directives unknown to
initial-setup, then initial-setup will not come up and fail silently.

Change-Id: Icffa48a88d15b07de70d729fcc59bf25e2e5d8ac
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>