Create a new appliance report which is smarter
Instead of passing the qcow, use an OVA

It will extract it, parse the OVF to get the actual values, then
use guestfish to get a list of key packages.

This requires ~2GB of free space in the folder you run it from,
and takes more memory, but is much more flexible. No more
hardcoding

Change-Id: Ibe2c4020d735a426f22ece9846c000436883e454
Signed-off-by: Ryan Barry <rbarry@redhat.com>