Install ovirt-provider-ovn on engine-appliance
Engine appliance must have ovirt-provider-ovn
preinstalled, in case the user wants to install
it in an offline engine-setup.

We do not want to have a hard dependency in the
form of an rpm require.
OVN and openvswitch are relatively heavy and complex,
and are still experimental. We would not want to
force everybody to pull them onto any Engine host.
However, we would like to enable them on the Engine appliance.

Change-Id: Icc792536b9655e4ec1123bc620d8fc7878ee692c
Signed-off-by: Marcin Mirecki<mmirecki@redhat.com>