answer file: Disable grafana setup
With remote dwh db upgrade, engine-setup prompts the user for grafana
DB credentials, but hosted-engine runs it noninteractively thus fails
it. So disable grafana for hosted-engine.

This is likely not a big regression, as larger hosted-engine setups are
expected anyway to separate dwh to its own machine, now with grafana.

Change-Id: I48aca34f31118ee1fb03ee2ba7a5fc3c850652ee
Bug-Url: https://bugzilla.redhat.com/1866780
Signed-off-by: Yedidyah Bar David <didi@redhat.com>