Really fail on any failed %post command
We included --erroronfail, but it seems to be not enough in case
of bash, since it jumps over any failed commands, thus it will
only fail if the last command in the block fails.

This is not the behaviour that we want, as we want to fail the
installation if any of the steps fail.

Signed-off-by: Lev Veyde <lveyde@redhat.com>